refactor(DOMRecFrame): convert prototype pattern to es6 class
