+++ src/constants.ts
@@ -1,0 +1,14 @@
+export const DOMREC_ADD = 'a';
+export const DOMREC_CANVAS_DATA = 'c';
+export const DOMREC_DELAY = 'd';
+export const DOMREC_FRAME = 'e';
+export const DOMREC_FORCE_STYLE_FLUSH = 'f';
+export const DOMREC_INPUT = 'i';
+export const DOMREC_LABEL = 'l';
+export const DOMREC_MOUSE_MOVE = 'm';
+export const DOMREC_MOUSE_DOWN = 'n';
+export const DOMREC_ATTR = 'r';
+export const DOMREC_TEXT = 't';
+export const DOMREC_MOUSE_UP = 'u';
+export const DOMREC_REMOVE = 'v';
+export const DOMREC_SCROLL = 's';
